refactor: ещё рефакторинг

+++ src/main/java/ru/yandex/practicum/filmorate/model/Likes.java
@@ -6,6 +6,6 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 public class Likes {
-    private int user_id;
-    private int film_id;
+    private Long user_id;
+    private Long film_id;
 }
